cleared up errors from folder renaming
